refactor(nx-set-shas): revamped nx-set-shas command
now is ussing commander 11 and @nx-tools/core for logging
