+++ package.json
@@ -20,8 +20,7 @@
     "@renovate/pep440": "1.0.0",
     "@swc/helpers": "~0.5.0",
     "ci-info": "3.8.0",
-    "colorette": "2.0.20",
-    "commander": "9.5.0",
+    "commander": "11.0.0",
     "csv-parse": "5.4.0",
     "handlebars": "4.7.7",
     "moment": "2.29.4",
