+++ packages/nx-set-shas/jest.config.ts
@@ -2,15 +2,9 @@
 export default {
   displayName: 'nx-set-shas',
   preset: '../../jest.preset.js',
-  globals: {},
   testEnvironment: 'node',
   transform: {
-    '^.+\\.[tj]s$': [
-      'ts-jest',
-      {
-        tsconfig: '<rootDir>/tsconfig.spec.json',
-      },
-    ],
+    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
   coverageDirectory: '../../coverage/packages/nx-set-shas',
