+++ packages/nx-set-shas/package.json
@@ -1,13 +1,16 @@
 {
   "name": "@nx-tools/nx-set-shas",
-  "version": "5.0.3",
+  "version": "6.0.0-alpha.1",
   "type": "commonjs",
   "bin": {
     "nx-set-shas": "./src/index.js"
   },
+  "main": "./src/index.js",
+  "typings": "./src/index.d.ts",
   "dependencies": {
-    "commander": "9.5.0",
-    "colorette": "2.0.20",
+    "@nx-tools/core": "6.0.0-alpha.1",
+    "chalk": "4.1.2",
+    "commander": "11.0.0",
     "undici": "5.23.0"
   },
   "peerDependencies": {
