+++ packages/nx-set-shas/project.json
@@ -11,16 +11,11 @@
         "outputPath": "dist/packages/nx-set-shas",
         "main": "packages/nx-set-shas/src/index.ts",
         "tsConfig": "packages/nx-set-shas/tsconfig.lib.json",
-        "assets": ["packages/nx-set-shas/*.md"],
-        "buildableProjectDepsInPackageJsonType": "dependencies",
-        "updateBuildableProjectDepsInPackageJson": true
+        "assets": ["packages/nx-set-shas/*.md"]
       }
     },
     "publish": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "node tools/scripts/publish.mjs nx-set-shas {args.ver} {args.tag}"
-      },
+      "command": "node tools/scripts/publish.mjs nx-set-shas {args.ver} {args.tag}",
       "dependsOn": ["build"]
     },
     "lint": {
@@ -36,6 +31,12 @@
       "options": {
         "jestConfig": "packages/nx-set-shas/jest.config.ts",
         "passWithNoTests": true
+      },
+      "configurations": {
+        "ci": {
+          "ci": true,
+          "codeCoverage": true
+        }
       }
     }
   },
